% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastime_constructor.R
\name{new_fastime_bridge}
\alias{new_fastime_bridge}
\title{Constructor for creating fastime models}
\usage{
new_fastime_bridge(class, models, data, extras = NULL, desc = NULL)
}
\arguments{
\item{class}{A class name that is used for creating custom printing messages}

\item{models}{A list containing one or more models}

\item{data}{A data.table containing 4 columns:
(date column with name that matches input data), .actual, .fitted, and .residuals.}

\item{extras}{An optional list.}

\item{desc}{An optional model description to appear when printing your fastime objects}
}
\description{
These functions are used to construct new \code{fastime} bridge functions
}
